Mirage-Project/backend-python/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python
requests
